refactor: Split webpack config file into dev and prod to get rid of inline conditionals
