+++ package.json
@@ -71,6 +71,7 @@
     "semantic-release": "^24.1.0",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.0.4"
+    "webpack-merge": "^6.0.1"
   },
   "scripts": {
     "start": "node server",
